Add checks for missing tools in the build script
If the required tools, i.e. dosfstools or sgdisk are missing, the
script will behave unpredictably, sometimes doing things it wasn't
supposed to do, i.e. try to call sgdisk (which is missing) and then
setting up a loop device based on that.

This fixes hard-to-debug errors and makes it easier to understand when
some tool is missing. This doesn't cover everything, because I'm not
sure what is needed, but should be good enough for most linux systems.